+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Features:
 
 - Add ``pre_load`` and ``post_load`` parameters to `marshmallow.fields.Field` for
   field-level pre- and post-processing (:issue:`2787`).
+- Typing: improvements to `marshmallow.validate` (:pr:`2940`).
 
 4.2.4 (2026-04-02)
 ------------------